requirements.txt
numpy==1.15.4
scipy==1.1.0
torch==1.0.0
wrapt==1.11.2
matplotlib==3.0.3
torchvision==0.2.1
cvxopt==1.2.3
circlify==0.10.0
Pillow==6.1.0
scikit_learn==0.21.3
